Pull up listener for MakeCode project load for code page
